Modify Order and artwork classes

+++ src/DTO/ArtworkDTO.cs
@@ -15,11 +15,10 @@ namespace Backend_Teamwork.src.DTO
             public string Title { get; set; }
             public string Description { get; set; }
             public int Quantity { get; set; }
-            public double Price { get; set; }
+            public decimal Price { get; set; }
             public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
             public Guid ArtistId { get; set; }
             public Guid CategoryId { get; set; }
-
         }
 
         // read data (get data)
